backend/requirements.txt
<<<<<<< HEAD
# =========================================================================
# Stock Scanner - Linux Requirements
# MySQL + yfinance + NASDAQ-only tickers
# =========================================================================

# ===== CORE DJANGO FRAMEWORK =====
Django>=4.2.11,<5.0
django-extensions>=3.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
whitenoise>=6.5.0
python-json-logger>=2.0.7
sentry-sdk>=2.16.0

# ===== DATABASE & CACHING =====
mysqlclient>=2.2.0
dj-database-url>=2.1.0

# ===== STOCK DATA & API =====
yfinance>=0.2.25
requests>=2.31.0
urllib3>=2.0.0

# ===== TASK QUEUE & SCHEDULING =====
celery>=5.3.0
schedule>=1.2.0

# ===== DATABASE SUPPORT =====
dj-database-url>=3.0.0
PyMySQL>=1.1.0
cryptography>=45.0.0
# psycopg2-binary>=2.9.0  # For PostgreSQL (commented out)

# ===== ENVIRONMENT & UTILITIES =====
python-dotenv>=1.0.0

# ===== TEXT PROCESSING & NLP =====
textblob>=0.17.1

# ===== NEWS SCRAPING =====
feedparser>=6.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ===== DATA PROCESSING (OPTIONAL) =====
# These are optional and will be installed if available
numpy>=1.24.0
pandas>=2.0.0

# ===== SECURITY & CRYPTO =====
cryptography>=41.0.0

# ===== MARKET HOURS MANAGER =====
pytz>=2023.3
psutil>=5.9.0
=======
fastapi==0.110.1
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
tzdata>=2024.2
motor==3.3.1
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.9
jq>=1.6.0
typer>=0.9.0
>>>>>>> b9dee287 (auto-commit for f45bf728-febb-4567-ac8e-02aafd409816)
